Sprouts/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
PyMuPDF>=1.23.0
openai>=1.0.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0 